fix(engine): read hash_bytes' tail byte as unsigned
`end[i]` is `char`, signed here. A tail byte >= 0x80 sign-extends to
0xFFFFFFFFFFFFFF00 | byte, and the `or` then sets every bit above bit 7 --
erasing everything already accumulated from the higher indices. Reference
MurmurHash64A reads through `const unsigned char*`.

Measured over the whole two-byte domain: 32896 distinct values of 65536.
{0x80,0x02,0x03} and {0x80,0x02,0x04} both hashed to 2e1c141ddbba99f3.

Not a dead path. It hashes NNUE layer content during plain startup, and the
value flows into Network::get_content_hash and then into the system-wide
shared-memory segment name, so two different nets could name one segment.

The check lives in tests/enginelink_main.cpp, which is the only program in
the tree that links engine/ alone and runs it, and it is stated as a domain
count rather than as a golden hash: a golden pins the algorithm, and what has
to hold is that distinct inputs get distinct hashes.

One consequence, stated rather than discovered: every value hash_bytes
produces changes, so a fixed and an unfixed binary name different
shared-memory segments and stop sharing with each other. Nothing corrupts --
each names its own. No hash reaches a file or a version boundary; the readers
are get_content_hash and hash_string, both live-process only.

No functional change

+++ src/engine/hashing.h
@@ -55,9 +55,15 @@ inline u64 hash_bytes(const char* data, usize size) {
 
     if (size & 7)
     {
+        // Read the tail byte as UNSIGNED. `end[i]` is `char`, which is signed
+        // here: a byte >= 0x80 sign-extends to 0xFFFFFFFFFFFFFF00 | byte, and
+        // the `or` then sets every bit above bit 7 -- erasing everything already
+        // accumulated from the higher indices. Reference MurmurHash64A reads
+        // through `const unsigned char*`. Over all 65536 two-byte inputs the
+        // signed form produces 32896 distinct values instead of 65536.
         u64 k = 0;
         for (int i = (size & 7) - 1; i >= 0; i--)
-            k = (k << 8) | u64(end[i]);
+            k = (k << 8) | u64(u8(end[i]));
 
         h ^= k;
         h *= m;
